Refactorized vault dashboard header
