+++ savewise-client/src/components/VaultDashboard/VaultDashboard.module.scss
@@ -1,17 +1,3 @@
-.vaultSelect {
-    width: 15rem;
-}
-
-.vaultDashboardHeader {
-    display: flex;
-    align-items: center;
-}
-
-.backButton {
-    width: 55px;
-    height: 55px;
-}
-
 .componentBody {
     height: 100%;
     display: flex;
