+++ savewise-client/src/components/VaultDashboardHeader/VaultDashboardHeader.module.scss
@@ -1,0 +1,13 @@
+.vaultDashboardHeader {
+    display: flex;
+    align-items: center;
+}
+
+.vaultSelect {
+    width: 15rem;
+}
+
+.backButton {
+    width: 55px;
+    height: 55px;
+}
